requirements.txt
google-cloud==0.34.0
google-cloud-storage==2.18.2
python-dotenv==1.1.0
requests==2.32.3
deepeval==3.0.3
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
asyncpg==0.30.0
fastapi[standard]==0.115.12
uvicorn==0.34.3
pydantic==2.11.5